+++ test/color-extractor-strategies/hexa-extrator.test.ts
@@ -68,3 +68,36 @@ describe('Test CSS hexa color Regex', () => {
     assert.notOk('#fffffff'.match(REGEXP));
   });
 });
+describe('Test CSS hexa (with 0x prefix) color Regex', () => {
+  it('Should match color with only integer', function () {
+    assert.ok('0x000000'.match(REGEXP));
+  });
+  it('Should match color with letters and integers', function () {
+    assert.ok('0xf0f0f0'.match(REGEXP));
+  });
+  it('Should match color with only letters', function () {
+    assert.ok('0xffffff'.match(REGEXP));
+  });
+  it('Regex should not care about the case', function () {
+    assert.ok('0xAbc012'.match(REGEXP));
+  });
+  it('Sould match inside a string', function() {
+    assert.ok('"0xffffff"'.match(REGEXP));
+  });
+  it('Should match with different characters at the end', function () {
+    assert.ok('0xffffff '.match(REGEXP));
+    assert.ok('0xffffff,'.match(REGEXP));
+    assert.ok('0xffffff;'.match(REGEXP));
+    assert.ok('0xffffff\n'.match(REGEXP));
+    assert.ok('0xffffff)'.match(REGEXP));
+  });
+  it('Should accept accept an alpha value', function() {
+    assert.ok('0x00000000'.match(REGEXP));
+  });
+  it('Should not match', function () {
+    assert.notOk('0xfffffg'.match(REGEXP));
+    assert.notOk('0xffffffg'.match(REGEXP));
+    assert.notOk('0xfffff'.match(REGEXP));
+    assert.notOk('0xfffffff'.match(REGEXP));
+  });
+});