+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.6.18
+
+- [FIX] Sass variables extraction
+
+## 0.6.17
+
+- [FIX] Colorize uppercased browsers colors (red, blue...)
+- [IMPROVE] Colorize hexa colors with 0x prefix
+- [IMPROVE] Update variables extractions. Split the variable extractions in specialized strategies (one for sass, css, less ...).
+
 ## 0.6.16
 
 - [FIX] Split the regexp extracting variables in three to prevent a crash (the previous one was too complex).